Add files via upload
We have made some very necessary changes to the code and updated it. It can clear the console if we run it on CMD or Powershell. Also, the grid view is updated. With this code, you can generate a random wordoku puzzle. We are working on the next edition of the project. Consider this code for this week.